object-based experiment interface
Major refactoring of PlanOut code.  The PlanOut interpreter and
PlanOutKit object are now all 'Mapper' objects, which have interfaces
to get parameters.  The Experiment abstract base class structures
auto-exposure logging. Documentation is provided for how the object can
be extended for use in production environments. A SimpleExperiment
class is used for testing / demos. It logs to a log file whose name is
the same as the experiment name. Some of the example experiments are
transitioned from the decorator syntax to the class syntax.