Fixed multiple logging bug with SimpleExperiment
If multiple instances were made of a SimpleExperiment, a logging
handler would be added multiple times, causing calls to logger to log
multiple times. This fixes this issue by using a class variable to make
sure the logger is only configured once. The file name of the log was
also modified to not turn the name into lower case, to make it a little
more transparent that the class name is used as the default experiment
name and log filename.