refactor PlanOutMappers to not rely on PlanOutInterpreterMapper
Currently PlanOutKitMapper requires a PlanOutInterpreterMapper object
to work.  This was mostly because PlanOutOpRandom read the experiment
salt from the interpreter environment. This makes it difficult to port
PlanOut without porting the entire interpreter.

This change makes the experiment salt an instance variable of each
PlanOutMapper so that the interpreter object doesn’t need to be used
for this purpose. Experiment-level salts are now set in the mapper’s
constructor, rather than via PlanOut code.  This change also cleans up
some cruft and makes abstract base classes abstract using abc.