improve interpreter
-made PlanOutInterpreterMapper take input data as a variable in the
constructor. this data gets stored under self._inputs. get() plays nice
with this, so that a PlanOut script being executed doesn’t know an
input variable isn’t part of its own internal environment. the upside
is that when used in conjunction with the experiment classes, inputs
will be logged separately from planout environment variables, so that
the log formats for analogous PlanOutInterpreter and PlanOutKit
experiments would be the same.
-Fixed a bug where arrays of operators would not be evaluated.

-Created a PlanOutInterpreterInspector class to handle inspection
(e.g., validation, pretty printing, and other future functionality).
This stuff is useful, and helps clarify why PlanOut scripts are nice,
but doesn’t belong as part of a Mapper class.

-Fixed bug with SetOverride
-Fixed bug where Arrays() weren’t being pretty printed correctly
-Fixed bug where * weren’t being rendered correctly
-Updated Sum to use PlanOutOpCommutative base class.
-Made salt not required for random ops (should be taken care of by set).

-started updating planout demo to handle new experiment class and
experiment inspection objects.