Added a namespace class which is used like an Experiment object. A SimpleNamespace object is included so that users can set up simple namespaces without using a database.
New classes include:
-Namespace: abstract class with core methods / properties all namespaces
should implement
-SimpleNamespace: abstract class implementing non-database backed
(PlanOut based) segment allocation. SimpleNamespace namespaces are
created by extending this class and implementing setup_attributes() and
setup_experiments().
-DefaultExperiment: an experiment class that has no logging and returns
a keystore/dictionary by default. This is intended for use with
Namespaces when units are not allocated to an active experiment.

Changes to Experiment class
-setup() has been changed to setup_attributes() to mimic namespaces

Changes to SimpleExperiment
-class-variable logger works with multiple experiment names. This makes
it possible to have multiple instances of the same experiment class
running with different names (as one may after deciding to more broadly
roll out an experiment to more users).