Prevent excessive logging to the console
This is a simple change that represents one way to make the IPython notebooks work for everyone when you demo it in the fashion used for PyData 2014. By default, all handled messages will be handled by the root logger as well, which is set by default to log to the console. It doesn't matter that the root logger is set to log only warnings.